fix main
