+++ main.py
@@ -1,63 +1,114 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Nov 10 10:42:39 2022
-
-@author: iannello
-
-"""
-
+import matplotlib.pyplot as plt
 from Scacchiera import Scacchiera
 from Torre import Torre
 from Alfiere import Alfiere
-from pedone import Pedone
 from Cavallo import Cavallo
+from pedone import Pedone  # aggiungi import pedone
 
+# Variabili globali per la selezione
+selezione = {'partenza': None, 'destinazione': None}
+scacchiera = None  # sarà inizializzata nel main
+
+def coord_da_click(event):
+    """Traduci coordinate matplotlib in coordinate scacchiera."""
+    if event.inaxes is None:
+        return None
+    col = int(round(event.xdata))
+    row = int(round(event.ydata))
+    if 0 <= col < 8 and 0 <= row < 8:
+        colonna = list('ABCDEFGH')[col]
+        riga = 8 - row
+        return [colonna, riga]
+    return None
+
+def on_click(event):
+    """Gestione click per selezionare e muovere pezzi."""
+    global selezione, scacchiera
+    pos = coord_da_click(event)
+    if pos is None:
+        return
+    if selezione['partenza'] is None:
+        pezzo = scacchiera.get_pezzo(pos)
+        if pezzo is not None:
+            selezione['partenza'] = pos
+            print(f"Selezionata partenza: {pos}")
+    else:
+        selezione['destinazione'] = pos
+        pezzo = scacchiera.get_pezzo(selezione['partenza'])
+        if pezzo and pezzo.verifica_mossa(selezione['destinazione']):
+            # --- EN PASSANT ---
+            if isinstance(pezzo, Pedone):
+                r0, c0 = selezione['partenza']
+                r1, c1 = selezione['destinazione']
+                if abs(c1 - c0) == 1 and r1 != r0:
+                    if scacchiera.get_pezzo(selezione['destinazione']) is None:
+                        casella_adiacente = [r0, c1]
+                        pedone_vicino = scacchiera.get_pezzo(casella_adiacente)
+                        if (
+                            pedone_vicino is not None
+                            and isinstance(pedone_vicino, Pedone)
+                            and pedone_vicino.colore != pezzo.colore
+                            and getattr(pedone_vicino, "ha_fatto_doppio_passo", False)
+                        ):
+                            print("Cattura en passant!")
+                            scacchiera.togli(casella_adiacente)
+            # --- FINE EN PASSANT ---
+
+            if scacchiera.get_pezzo(selezione['destinazione']) is not None:
+                scacchiera.togli(selezione['destinazione'])
+            scacchiera.togli(selezione['partenza'])
+            scacchiera.metti(pezzo, selezione['destinazione'])
+
+            # --- PROMOZIONE PEDONE ---
+            if isinstance(pezzo, Pedone):
+                r, c = selezione['destinazione']
+                if (pezzo.colore == "W" and r == 'H') or (pezzo.colore == "B" and r == 'A'):
+                    print("Promozione del pedone!")
+                    while True:
+                        scelta = input("Promuovi in (Q,R,B,N): ").upper()
+                        if scelta in {"Q", "R", "B", "N"}:
+                            break
+                        print("Scelta non valida.")
+                    nuovo = pezzo.promuovi(scelta)
+                    scacchiera.togli(selezione['destinazione'])
+                    scacchiera.metti(nuovo, selezione['destinazione'])
+            # --- FINE PROMOZIONE ---
+
+            print(f"Mossa: {selezione['partenza']} -> {selezione['destinazione']}")
+        else:
+            print("Mossa non valida")
+        selezione = {'partenza': None, 'destinazione': None}
+        plt.close()  # Chiudi la finestra per ridisegnare
+        scacchiera.visualizza(on_click=on_click)
+
+# Funzioni per posizionare i pezzi
 def metti_alfiere(scacchiera: Scacchiera):
     #bianchi
-    alfiere1= Alfiere("W")
-    scacchiera.metti(alfiere1, ['H', 3])
-    alfiere2= Alfiere("W")
-    scacchiera.metti(alfiere2, ['H', 6])
-
+    scacchiera.metti(Alfiere("W"), ['H', 3])
+    scacchiera.metti(Alfiere("W"), ['H', 6])
     #neri
-    alfiere3= Alfiere("B")
-    scacchiera.metti(alfiere3, ['A', 3])
-    alfiere4= Alfiere("B")
-    scacchiera.metti(alfiere4, ['A', 6])
+    scacchiera.metti(Alfiere("B"), ['A', 3])
+    scacchiera.metti(Alfiere("B"), ['A', 6])
 
 def metti_torre(scacchiera: Scacchiera):
     #bianchi
-    torre1= Torre("W")
-    scacchiera.metti(torre1, ['H', 1])
-    torre2= Torre("W")
-    scacchiera.metti(torre2, ['H', 8])
-
+    scacchiera.metti(Torre("W"), ['H', 1])
+    scacchiera.metti(Torre("W"), ['H', 8])
     #neri
-    torre3= Torre("B")
-    scacchiera.metti(torre3, ['A', 1])
-    
-    torre4= Torre("B")
-    scacchiera.metti(torre4, ['A', 8])
+    scacchiera.metti(Torre("B"), ['A', 1])
+    scacchiera.metti(Torre("B"), ['A', 8])
 
 def metti_cavallo(scacchiera: Scacchiera):
-    # bianchi
-    cavallo1 = Cavallo("W")
-    scacchiera.metti(cavallo1, ['H', 2])
-    cavallo2 = Cavallo("W")
-    scacchiera.metti(cavallo2, ['H', 7])
+    #bianchi
+    scacchiera.metti(Cavallo("W"), ['H', 2])
+    scacchiera.metti(Cavallo("W"), ['H', 7])
+    #neri
+    scacchiera.metti(Cavallo("B"), ['A', 2])
+    scacchiera.metti(Cavallo("B"), ['A', 7])
 
-    # neri
-    cavallo3 = Cavallo("B")
-    scacchiera.metti(cavallo3, ['A', 2])
-    cavallo4 = Cavallo("B")
-    scacchiera.metti(cavallo4, ['A', 7])
-
-
-def metti_pedone(scacchiera: Scacchiera):
+def metti_pedone(schacchiera: Scacchiera):
     for col in range(1, 9):
         scacchiera.metti(Pedone("W"), ['B', col])
-
     for col in range(1, 9):
         scacchiera.metti(Pedone("B"), ['G', col])
 
@@ -67,133 +118,17 @@ def metti_re():
 def metti_regina():
     pass
 
-
-def in_board(posizione):
-    """
-    verifica che la posizione sia all'intgerno della scacchiera
-    Parameters
-    ----------
-    posizione: coppia di coordinate
-
-    Returns
-    -------
-    bool
-        True se le coordinate corrispondono a una casella della
-        scacchiera, False altrimenti
-    """
-    return posizione[0] in {'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H'} and \
-           posizione[1] in range(1, 9)
-
-
-def get_mossa():
-    """
-    acquisisce una mossa dallo standard input o termina il programma
-    La mossa deve essere fornita nel formato:
-        
-        posizione_di_partenza posizione_di_destinazione
-        
-    dove una posizione è una coppia formata da una lettera
-    in ['A', 'H'] e da una cifra in [1, 8]
-    le due posizioni devono essere separate da un solo spazio
-
-    se la lunghezza della stringa fornita in input è diversa
-    da 5 il programma viene terminato
-
-    Returns
-    -------
-    list
-        posizione di partenza
-    list
-        posizione di destinazione
-
-    """
-    while True:
-        mossa = input("Dammi la mossa: ")
-        if not len(mossa) == 5:  # l'input non è una mossa
-            exit(0)              # termina il programma
-        partenza = [mossa[0].upper(), int(mossa[1])]
-        destinazione = [mossa[3].upper(), int(mossa[4])]
-        if in_board(partenza) and in_board(destinazione):
-            return partenza, destinazione
-        else:
-            print(f'La partenza e/o la destinazione della mossa {mossa} non corrispondono a caselle della scacchiera')
-
 #TODO: gestire la mano
 #TODO: fare le directory ordinate
 #TODO: scrivere commenti
 if __name__ == "__main__":
-    # setup del gioco
     scacchiera = Scacchiera()
-    # posizione 4 pezzi bianchi nelle prime 4 righe della colonna A
-    
     metti_alfiere(scacchiera)
-    
     metti_torre(scacchiera)
-    
     metti_pedone(scacchiera)
 
     #metti_re(scacchiera)
 
     #metti_regina(scacchiera)
-
     metti_cavallo(scacchiera)
-
-    scacchiera.visualizza()
-    print()
-
-    # inizia il gioco
-    while True:
-        while True:
-            # acquisisce mossa da fare
-            (partenza, destinazione) = get_mossa()
-            # recupera il pezzo da muovere
-            pezzo = scacchiera.get_pezzo(partenza)
-            if pezzo is None:  # la mossa parte da una casella vuota
-                print(f'La mossa non è valida: la casella di partenza è vuota')
-            elif pezzo.verifica_mossa(destinazione):  # la mossa è legale
-                break
-
-        # 2) en passant
-        if isinstance(pezzo, Pedone):
-            r0, c0 = partenza
-            r1, c1 = destinazione
-            if abs(c1 - c0) == 1 and r1 != r0:
-                if scacchiera.get_pezzo(destinazione) is None:
-                    casella_adiacente = [r0, c1]
-                    pedone_vicino = scacchiera.get_pezzo(casella_adiacente)
-                    if (
-                        pedone_vicino is not None
-                        and isinstance(pedone_vicino, Pedone)
-                        and pedone_vicino.colore != pezzo.colore
-                        and pedone_vicino.ha_fatto_doppio_passo
-                    ):
-                        print("Cattura en passant!")
-                        scacchiera.togli(casella_adiacente)
-
-        # esegui mossa sulla scacchiera
-        if not scacchiera.get_pezzo(destinazione) is None:  # la casella è occupata
-            scacchiera.togli(destinazione)  # "mangia" il pezzo che occupa la casella
-        scacchiera.togli(partenza)
-        scacchiera.metti(pezzo, destinazione)
-
-        # 5) PROMOZIONE
-        if isinstance(pezzo, Pedone):
-            r, c = destinazione
-
-            if (pezzo.colore == "W" and r == 'H') or (pezzo.colore == "B" and r == 'A'):
-
-                print("Promozione del pedone!")
-
-                while True:
-                    scelta = input("Promuovi in (Q,R,B,N): ").upper()
-                    if scelta in {"Q", "R", "B", "N"}:
-                        break
-                    print("Scelta non valida.")
-
-                nuovo = pezzo.promuovi(scelta)
-                scacchiera.togli(destinazione)
-                scacchiera.metti(nuovo, destinazione)
-
-        scacchiera.visualizza()
-        print()
-
+    scacchiera.visualizza(on_click=on_click)
